+++ swcgeom/utils/transforms.py
@@ -27,7 +27,7 @@ def angle(a: npt.ArrayLike, b: npt.ArrayLike) -> float:
     return theta if np.cross(a, b) > 0 else -theta
 
 
-def scale3d(sx: float, sy: float, sz: float) -> npt.NDArray[np.float64]:
+def scale3d(sx: float, sy: float, sz: float) -> npt.NDArray[np.float32]:
     """Get the 3D scale transfomation matrix.
 
     Returns
@@ -42,11 +42,11 @@ def scale3d(sx: float, sy: float, sz: float) -> npt.NDArray[np.float64]:
             [0, 0, sz, 0],
             [0, 0, 0, 1],
         ],
-        dtype=np.float64,
+        dtype=np.float32,
     )
 
 
-def translate3d(tx: float, ty: float, tz: float) -> npt.NDArray[np.float64]:
+def translate3d(tx: float, ty: float, tz: float) -> npt.NDArray[np.float32]:
     """Get the 3D translate transfomation matrix.
 
     Returns
@@ -61,11 +61,11 @@ def translate3d(tx: float, ty: float, tz: float) -> npt.NDArray[np.float64]:
             [0, 0, 1, tz],
             [0, 0, 0, 1],
         ],
-        dtype=np.float64,
+        dtype=np.float32,
     )
 
 
-def rotate3d(n: npt.ArrayLike, theta: float) -> npt.NDArray[np.float64]:
+def rotate3d(n: npt.ArrayLike, theta: float) -> npt.NDArray[np.float32]:
     r"""Get the 3D rotation transfomation matrix.
     
     Rotate v with axis n by an angle theta according to the right hand rule, 
@@ -104,7 +104,7 @@ def rotate3d(n: npt.ArrayLike, theta: float) -> npt.NDArray[np.float64]:
             [nz, 0, -nx],
             [-ny, nx, 0],
         ],
-        dtype=np.float64,
+        dtype=np.float32,
     )
 
     return (
@@ -114,7 +114,7 @@ def rotate3d(n: npt.ArrayLike, theta: float) -> npt.NDArray[np.float64]:
     )
 
 
-def rotate3d_x(theta: float) -> npt.NDArray[np.float64]:
+def rotate3d_x(theta: float) -> npt.NDArray[np.float32]:
     """Get the 3D rotation transfomation matrix.
 
     Rotate 3D vector `v` with `x`-axis by an angle theta according to the right
@@ -138,11 +138,11 @@ def rotate3d_x(theta: float) -> npt.NDArray[np.float64]:
             [0, np.sin(theta), np.cos(theta), 0],
             [0, 0, 0, 1],
         ],
-        dtype=np.float64,
+        dtype=np.float32,
     )
 
 
-def rotate3d_y(theta: float) -> npt.NDArray[np.float64]:
+def rotate3d_y(theta: float) -> npt.NDArray[np.float32]:
     """Get the 3D rotation transfomation matrix.
 
     Rotate 3D vector `v` with `y`-axis by an angle theta according to the right
@@ -165,11 +165,11 @@ def rotate3d_y(theta: float) -> npt.NDArray[np.float64]:
             [-np.sin(theta), 0, np.cos(theta), 0],
             [0, 0, 0, 1],
         ],
-        dtype=np.float64,
+        dtype=np.float32,
     )
 
 
-def rotate3d_z(theta: float) -> npt.NDArray[np.float64]:
+def rotate3d_z(theta: float) -> npt.NDArray[np.float32]:
     """Get the 3D rotation transfomation matrix.
 
     Rotate 3D vector `v` with `z`-axis by an angle theta according to the right
@@ -192,5 +192,5 @@ def rotate3d_z(theta: float) -> npt.NDArray[np.float64]:
             [0, 0, 1, 0],
             [0, 0, 0, 1],
         ],
-        dtype=np.float64,
+        dtype=np.float32,
     )
